This is a modified version of the program that is included in the VS Code C++ Extension.
The original code gets an error related to the syntax used in the for loop. This version is based on the W3Schools C++ Exercise showing how to loop through an array of string objects. The Notes at the bottom of the program are based in part from the stack overflow post on the sizeof() function: https://stackoverflow.com/questions/14308935/finding-the-size-of-an-array-of-strings-in-c